+++ src/compression.rs
@@ -4,7 +4,7 @@ use slow5lib_sys::{
 };
 
 /// SLOW5 record compression
-#[derive(Debug)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum RecordCompression {
     /// No compression
     None,
@@ -24,8 +24,20 @@ impl RecordCompression {
     }
 }
 
+impl From<u32> for RecordCompression {
+    #[allow(non_upper_case_globals)]
+    fn from(n: u32) -> Self {
+        match n {
+            slow5_press_method_SLOW5_COMPRESS_NONE => Self::None,
+            slow5_press_method_SLOW5_COMPRESS_ZLIB => Self::Zlib,
+            slow5_press_method_SLOW5_COMPRESS_ZSTD => Self::ZStd,
+            _ => unreachable!("Invalid record compression"),
+        }
+    }
+}
+
 /// SLOW5 signal compression
-#[derive(Debug)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum SignalCompression {
     /// No signal compression
     None,
@@ -38,6 +50,17 @@ impl SignalCompression {
         match self {
             SignalCompression::None => slow5_press_method_SLOW5_COMPRESS_NONE,
             SignalCompression::StreamVByte => slow5_press_method_SLOW5_COMPRESS_SVB_ZD,
+        }
+    }
+}
+
+impl From<u32> for SignalCompression {
+    #[allow(non_upper_case_globals)]
+    fn from(n: u32) -> Self {
+        match n {
+            slow5_press_method_SLOW5_COMPRESS_NONE => SignalCompression::None,
+            slow5_press_method_SLOW5_COMPRESS_SVB_ZD => SignalCompression::StreamVByte,
+            _ => unreachable!("Invalid compression"),
         }
     }
 }