+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-## [0.9.0] - yyyy-mm-dd
+## [0.9.0] - 2022-12-10
 
 ### Added